+++ cmd/bosc/run_tests.sh
@@ -13,51 +13,64 @@ rm string.bo
 #set -x
 rm tests/*.bos.o tests/*.bos.bo tests/*.bs tests/*.out tests/*.stdout
 
+fail=''
 
 for t in `ls tests/*_test.bos`; do
     echo -e "\n\n############################ $t ############################"
     ./bosc -o ${t}.bs $t >${t}.bosc.out 2>&1
     if [[ $? != 0 ]]; then
-	echo compiler failed for ${t}:
-	cat ${t}.bosc.out
-	echo ${t} FAIL
-	continue
+		echo compiler failed for ${t}:
+		cat ${t}.bosc.out
+		echo ${t} FAIL
+		fail=$(echo -e "$fail\n${t} FAIL")
+		continue
     fi
     ./bas -o ${t}.bo init_linux.bs ${t}.bs >${t}.bas.out 2>&1
     if [[ $? != 0 ]]; then
-	echo assembler failed for ${t}:
-	cat ${t}.bas.out
-	echo ${t} FAIL
-	continue
+		echo assembler failed for ${t}:
+		cat ${t}.bas.out
+		echo ${t} FAIL
+		fail=$(echo -e "$fail\n${t} FAIL")
+		continue
     fi
     # cat ${t}.bas.out
     ./bld -o ${t}.o ${t}.bo string.bo >${t}.bld.out 2>&1
     if [[ $? != 0 ]]; then
-	echo linker failed for ${t}:
-	cat ${t}.bld.out
-	echo ${t} FAIL
-	continue
+		echo linker failed for ${t}:
+		cat ${t}.bld.out
+		echo ${t} FAIL
+		fail=$(echo -e "$fail\n${t} FAIL")
+		continue
     fi
     ${t}.o > ${t}.stdout
     ecode="$?"
     if [[ "$ecode" != "0" ]]; then
-	echo $t exited with $ecode
-	echo -e 'stdout:\n```'
-	cat ${t}.stdout
-	echo '```'
-	echo ${t} FAIL
-	continue
+		echo $t exited with $ecode
+		echo -e 'stdout:\n```'
+		cat ${t}.stdout
+		echo '```'
+		echo ${t} FAIL
+		fail=$(echo -e "$fail\n${t} FAIL")
+		continue
     fi
     if [[ -f "${t}.expected" ]]; then
-	#echo "Comparing output."
-	diff -u "${t}.expected" "${t}.stdout"
-	if [[ $? != 0 ]]; then
-	    echo ${t} FAIL
-	    continue
-	fi
+		#echo "Comparing output."
+		diff -u "${t}.expected" "${t}.stdout"
+		if [[ $? != 0 ]]; then
+			echo ${t} FAIL
+			fail=$(echo -e "$fail\n${t} FAIL")
+			continue
+		fi
     fi
     echo ${t} PASS
 done
 
+if [[ $fail != '' ]]; then
+	echo -e '\nSUITE FAILED:'
+	echo -e "${fail}"
+	exit 1
+fi
 
 rm tests/*.bos.o tests/*.bos.bo tests/*.bs tests/*.out tests/*.stdout
+echo "SUITE PASSED"
+exit 0
